+++ src/components/Player.svelte
@@ -56,7 +56,22 @@
   }
 </script>
 
-<div class="player" class:stopped={$stopped} on:mousedown|self|preventDefault class:dev={isDev}>
+<!-- tabindex="0" causes the on:focusin event to fire -->
+<!-- svelte-ignore a11y-no-noninteractive-tabindex -->
+<div
+  class="player"
+  class:stopped={$stopped}
+  class:dev={isDev}
+  tabindex="0"
+  on:focusin={(e) => {
+    console.log(e.relatedTarget, e.currentTarget)
+    if (e.relatedTarget instanceof HTMLElement) {
+      e.relatedTarget.focus()
+    } else {
+      e.currentTarget.blur()
+    }
+  }}
+>
   <div class="left">
     {#if !$stopped && $playingTrack}
       <div
@@ -86,11 +101,10 @@
     {/if}
   </div>
   <div class="middle">
-    <div class="controls" on:mousedown|preventDefault>
+    <div class="controls">
       <button
         class="side-controls shuffle"
         class:on={$shuffle}
-        tabindex="-1"
         on:click={() => ($shuffle = !$shuffle)}
       >
         <svg xmlns="http://www.w3.org/2000/svg" width="24" viewBox="-8 -8 40 40" height="24"
@@ -99,7 +113,7 @@
           /></svg
         >
       </button>
-      <button tabindex="-1" on:click={previous}>
+      <button on:click={previous}>
         <svg xmlns="http://www.w3.org/2000/svg" height="24" viewBox="0 0 24 24" width="24">
           <path
             d="M7 6c.55 0 1 .45 1 1v10c0 .55-.45 1-1 1s-1-.45-1-1V7c0-.55.45-1 1-1zm3.66 6.82l5.77 4.07c.66.47 1.58-.01 1.58-.82V7.93c0-.81-.91-1.28-1.58-.82l-5.77 4.07c-.57.4-.57 1.24 0 1.64z"
@@ -108,7 +122,7 @@
       </button>
 
       {#if $paused}
-        <button class="play" tabindex="-1" on:click={playPause}>
+        <button class="play" on:click={playPause}>
           <svg xmlns="http://www.w3.org/2000/svg" height="36" viewBox="0 0 24 24" width="36">
             <path
               d="M8 6.82v10.36c0 .79.87 1.27 1.54.84l8.14-5.18c.62-.39.62-1.29 0-1.69L9.54 5.98C8.87 5.55 8 6.03 8 6.82z"
@@ -116,7 +130,7 @@
           </svg>
         </button>
       {:else}
-        <button class="pause" tabindex="-1" on:click={playPause}>
+        <button class="pause" on:click={playPause}>
           <svg xmlns="http://www.w3.org/2000/svg" height="36" viewBox="0 0 24 24" width="36">
             <path
               d="M8 19c1.1 0 2-.9 2-2V7c0-1.1-.9-2-2-2s-2 .9-2 2v10c0 1.1.9 2 2 2zm6-12v10c0 1.1.9 2 2 2s2-.9 2-2V7c0-1.1-.9-2-2-2s-2 .9-2 2z"
@@ -125,7 +139,7 @@
         </button>
       {/if}
 
-      <button tabindex="-1" on:click={next}>
+      <button on:click={next}>
         <svg xmlns="http://www.w3.org/2000/svg" height="24" viewBox="0 0 24 24" width="24">
           <path
             d="M7.58 16.89l5.77-4.07c.56-.4.56-1.24 0-1.63L7.58 7.11C6.91 6.65 6 7.12 6 7.93v8.14c0 .81.91 1.28 1.58.82zM16 7v10c0 .55.45 1 1 1s1-.45 1-1V7c0-.55-.45-1-1-1s-1 .45-1 1z"
@@ -133,12 +147,7 @@
         </svg>
       </button>
 
-      <button
-        class="side-controls repeat"
-        class:on={$repeat}
-        tabindex="-1"
-        on:click={() => ($repeat = !$repeat)}
-      >
+      <button class="side-controls repeat" class:on={$repeat} on:click={() => ($repeat = !$repeat)}>
         <svg xmlns="http://www.w3.org/2000/svg" width="24" height="24" viewBox="-8 -8 40 40"
           ><path
             d="M2 12c0 .999.381 1.902.989 2.604l-1.098.732-.587.392c-.814-1.025-1.304-2.318-1.304-3.728 0-3.313 2.687-6 6-6h9v-3l6 4-6 4v-3h-9c-2.206 0-4 1.794-4 4zm20.696-3.728l-.587.392-1.098.732c.608.702.989 1.605.989 2.604 0 2.206-1.795 4-4 4h-9v-3l-6 4 6 4v-3h9c3.313 0 6-2.687 6-6 0-1.41-.489-2.703-1.304-3.728z"
@@ -151,7 +160,6 @@
       <input
         type="range"
         class="time"
-        tabindex="-1"
         min="0"
         max={sliderSteps}
         bind:value={sliderValue}
@@ -162,7 +170,7 @@
     </div>
   </div>
   <div class="right">
-    <button class="volume-icon" tabindex="-1" on:click={volume.toggle}>
+    <button class="volume-icon" on:click={volume.toggle}>
       {#if $volume > 0.5}
         <svg
           class="high"
@@ -204,15 +212,7 @@
         </svg>
       {/if}
     </button>
-    <input
-      type="range"
-      class="volume"
-      tabindex="-1"
-      min="0"
-      max="1"
-      step="0.01"
-      bind:value={$volume}
-    />
+    <input type="range" class="volume" min="0" max="1" step="0.01" bind:value={$volume} />
     <button class="queue" tabindex="-1" on:click={toggleQueueVisibility} class:on={$queueVisible}>
       <svg xmlns="http://www.w3.org/2000/svg" height="24px" viewBox="0 0 24 24" width="24px">
         <path
@@ -300,6 +300,7 @@
   .controls
     display: flex
     justify-content: center
+    align-items: center
   .time-bar
     display: flex
     align-items: center